old/requirements_conda.txt
absl-py==2.4.0
attrs==25.4.0
cffi==2.0.0
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.61.1
jax==0.7.1
jaxlib==0.7.1
kiwisolver==1.4.9
matplotlib==3.10.8
mediapipe==0.10.18
ml_dtypes==0.5.4
mpu6050-raspberrypi==1.2
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pillow==12.1.1
protobuf==4.25.8
PyAudio==0.2.14
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scipy==1.17.0
sentencepiece==0.2.1
six==1.17.0
smbus==1.1.post2
smbus2==0.6.0
sounddevice==0.5.5
